requirements.txt
astroid==1.6.5
colorama==0.3.9
Django==2.0.6
django-user-agents==0.3.2
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pycryptodomex==3.6.3
pylint==1.9.2
pytz==2018.4
PyYAML==3.12
six==1.11.0
ua-parser==0.8.0
user-agents==1.1.0
wrapt==1.10.11
